requirements/base.txt
Django==2.1.4
django-cors-headers==2.4.0
django-dotenv==1.4.2
django-environ==0.4.5
django-filter==2.0.0
django-sendgrid-v5==0.7.0
djangorestframework-jwt==1.11.0
djangorestframework==3.9.0
pillow==5.3.0
psycopg2==2.7.6.1
simple-crypt==4.1.7
